Changed structure of preset manager Now it has single source of truth which is X-macro
- Changed PresetConfig structure fields to corresponding JSON fields
- Added ASan/UBSan option to CMake
- Changed limits on edits/spin boxes

+++ include/mainwindow.h
@@ -21,7 +21,7 @@ public:
 
 private:
   MouseButton getMouseButton() const;
-  size_t getIntervalMs() const;
+  int getIntervalMs() const;
 
   void applySettings();
   void applyPreset(const PresetConfig& config);
